Refined the Output using the MLP-algorithm
type: I have refined the code for the MLP-algorithm and boosted my score on the Kaggle competition to 0.77511

body: I have implemented a pipeline and gridsearched for the best hyper parameters for my predicitions wiht the MLP-algorithm. With this change I made it into the top 51%. Next up I will try to engineer some features with the name column to improve my score.